requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
Babel==2.9.1
bestconfig==1.3.5
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
frozenlist==1.3.3
future==0.18.2
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.3

pycparser==2.21
pyTelegramBotAPI==4.7.1
python-decouple==3.6
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.2
